Move github/data to vcs/data
