+++ modules/core/src/main/scala/org/scalasteward/core/application/Config.scala
@@ -21,7 +21,7 @@ import cats.effect.Sync
 import org.http4s.Uri
 import org.scalasteward.core.application.Cli.EnvVar
 import org.scalasteward.core.git.Author
-import org.scalasteward.core.github.data.AuthenticatedUser
+import org.scalasteward.core.vcs.data.AuthenticatedUser
 import org.scalasteward.core.util
 import scala.sys.process.Process
