+++ modules/core/src/main/scala/org/scalasteward/core/io/WorkspaceAlg.scala
@@ -21,7 +21,7 @@ import cats.FlatMap
 import cats.implicits._
 import io.chrisdavenport.log4cats.Logger
 import org.scalasteward.core.application.Config
-import org.scalasteward.core.github.data.Repo
+import org.scalasteward.core.vcs.data.Repo
 
 trait WorkspaceAlg[F[_]] {
   def cleanWorkspace: F[Unit]
