+++ modules/core/src/main/scala/org/scalasteward/core/nurture/NurtureAlg.scala
@@ -20,7 +20,7 @@ import cats.implicits._
 import io.chrisdavenport.log4cats.Logger
 import org.scalasteward.core.application.Config
 import org.scalasteward.core.git.{Branch, GitAlg}
-import org.scalasteward.core.github.data.{NewPullRequestData, Repo}
+import org.scalasteward.core.vcs.data.{NewPullRequestData, Repo}
 import org.scalasteward.core.github.{GitHubApiAlg, GitHubRepoAlg}
 import org.scalasteward.core.model.Update
 import org.scalasteward.core.repoconfig.RepoConfigAlg
