+++ modules/core/src/main/scala/org/scalasteward/core/nurture/UpdateData.scala
@@ -17,7 +17,7 @@
 package org.scalasteward.core.nurture
 
 import org.scalasteward.core.git.{Branch, Sha1}
-import org.scalasteward.core.github.data.Repo
+import org.scalasteward.core.vcs.data.Repo
 import org.scalasteward.core.model.Update
 
 final case class UpdateData(
