+++ modules/core/src/main/scala/org/scalasteward/core/steward.scala
@@ -21,7 +21,7 @@ import cats.FlatMap
 import cats.effect.{ExitCode, IO, IOApp, Sync}
 import cats.implicits._
 import org.scalasteward.core.application.Context
-import org.scalasteward.core.github.data.Repo
+import org.scalasteward.core.vcs.data.Repo
 
 object steward extends IOApp {
   override def run(args: List[String]): IO[ExitCode] =
