+++ modules/core/src/main/scala/org/scalasteward/core/update/UpdateService.scala
@@ -20,8 +20,8 @@ import cats.implicits._
 import io.chrisdavenport.log4cats.Logger
 import org.scalasteward.core.dependency.{Dependency, DependencyRepository}
 import org.scalasteward.core.git.Sha1
-import org.scalasteward.core.github.data.PullRequestState.Closed
-import org.scalasteward.core.github.data.Repo
+import org.scalasteward.core.vcs.data.PullRequestState.Closed
+import org.scalasteward.core.vcs.data.Repo
 import org.scalasteward.core.model.Update
 import org.scalasteward.core.nurture.PullRequestRepository
 import org.scalasteward.core.sbt._
