+++ modules/core/src/main/scala/org/scalasteward/core/vcs/data/BranchOut.scala
@@ -14,28 +14,18 @@
  * limitations under the License.
  */
 
-package org.scalasteward.core.github.data
+package org.scalasteward.core.vcs.data
 
-import cats.implicits._
 import io.circe.Decoder
 import io.circe.generic.semiauto._
-import org.http4s.Uri
-import org.scalasteward.core.github.data.PullRequestState.Closed
-import org.scalasteward.core.util.uri.uriDecoder
+import org.scalasteward.core.git.Branch
 
-final case class PullRequestOut(
-    html_url: Uri,
-    state: PullRequestState,
-    title: String
-) {
-  def isClosed: Boolean =
-    state === Closed
-}
+final case class BranchOut(
+    name: Branch,
+    commit: CommitOut
+)
 
-object PullRequestOut {
-  implicit val pullRequestOutDecoder: Decoder[PullRequestOut] =
+object BranchOut {
+  implicit val branchOutDecoder: Decoder[BranchOut] =
     deriveDecoder
-
-  // prevent IntelliJ from removing the import of uriDecoder
-  locally(uriDecoder)
 }
