+++ modules/core/src/test/scala/org/scalasteward/core/git/GitAlgTest.scala
@@ -1,7 +1,7 @@
 package org.scalasteward.core.git
 
 import org.http4s.Uri
-import org.scalasteward.core.github.data.Repo
+import org.scalasteward.core.vcs.data.Repo
 import org.scalasteward.core.mock.MockContext._
 import org.scalasteward.core.mock.MockState
 import org.scalatest.{FunSuite, Matchers}
