+++ modules/core/src/test/scala/org/scalasteward/core/github/http4s/authenticationTest.scala
@@ -3,7 +3,7 @@ package org.scalasteward.core.github.http4s
 import cats.Id
 import org.http4s.headers.{Accept, Authorization}
 import org.http4s.{BasicCredentials, Headers, MediaType, Request}
-import org.scalasteward.core.github.data.AuthenticatedUser
+import org.scalasteward.core.vcs.data.AuthenticatedUser
 import org.scalatest.{FunSuite, Matchers}
 
 class authenticationTest extends FunSuite with Matchers {
