+++ modules/core/src/test/scala/org/scalasteward/core/vcs/data/BranchOutTest.scala
@@ -1,4 +1,4 @@
-package org.scalasteward.core.github.data
+package org.scalasteward.core.vcs.data
 
 import io.circe.parser
 import org.scalasteward.core.git.Sha1.HexString
