+++ modules/core/src/test/scala/org/scalasteward/core/vcs/data/PullRequestStateTest.scala
@@ -1,8 +1,8 @@
-package org.scalasteward.core.github.data
+package org.scalasteward.core.vcs.data
 
 import io.circe.Decoder
 import io.circe.syntax._
-import org.scalasteward.core.github.data.PullRequestState.{Closed, Open}
+import org.scalasteward.core.vcs.data.PullRequestState.{Closed, Open}
 import org.scalatest.{Assertion, FunSuite, Matchers}
 
 class PullRequestStateTest extends FunSuite with Matchers {
